fix(pandas): Suppress datetime parsing warnings and improve performance
- Add format detection to avoid pandas warning spam in production logs
- Extract detect_datetime_format to new superset/utils/pandas.py module
- Suppress warnings when format cannot be detected (mixed formats)
- Improve performance by 5x for consistent date formats
- Add comprehensive tests for warning suppression and format detection
